+++ tests/Unit/Drivers/Imagick/DriverTest.php
@@ -59,10 +59,10 @@ final class DriverTest extends BaseTestCase
     /**
      * @param array<string|DecoderInterface> $decoders
      */
-    #[DataProviderExternal(InputDataProvider::class, 'handleImageInputDataProvider')]
+    #[DataProviderExternal(InputDataProvider::class, 'decodeImageDataProvider')]
     public function testHandleImageInput(mixed $input, ?array $decoders, string $resultClassname): void
     {
-        $this->assertInstanceOf($resultClassname, $this->driver->handleImageInput($input, $decoders));
+        $this->assertInstanceOf($resultClassname, $this->driver->decodeImage($input, $decoders));
     }
 
     /**
@@ -77,7 +77,7 @@ final class DriverTest extends BaseTestCase
     /**
      * @param array<string|DecoderInterface> $decoders
      */
-    #[DataProviderExternal(InputDataProvider::class, 'handleImageInputDataProvider')]
+    #[DataProviderExternal(InputDataProvider::class, 'decodeImageDataProvider')]
     public function testHandleColorInputFail(mixed $input, ?array $decoders, string $resultClassname): void
     {
         $this->expectException(ImageException::class);
@@ -91,7 +91,7 @@ final class DriverTest extends BaseTestCase
     public function testHandleImageInputFail(mixed $input, ?array $decoders, string $resultClassname): void
     {
         $this->expectException(ImageException::class);
-        $this->driver->handleImageInput($input);
+        $this->driver->decodeImage($input);
     }
 
     public function testColorProcessor(): void